# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for the Heracle personal training & fitness platform.
  title: Heracle Fitness API
  version: "1.0"
paths:
  /auth/admin/login:
    post:
      consumes:
      - application/json
      description: Authenticates admin with username and password
      parameters:
      - description: Admin credentials
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Admin user and JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
      summary: Admin login
      tags:
      - Auth
  /auth/dev/token:
    get:
      description: Creates or retrieves a user by email and returns a JWT (dev only)
      parameters:
      - default: sanjaysagar.main@gmail.com
        description: Email address
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User object and JWT token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      summary: Get development token
      tags:
      - Auth
  /auth/google/token:
    post:
      consumes:
      - application/json
      description: Verifies a Firebase ID token and creates or retrieves the user
      parameters:
      - description: Google auth payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: User object and JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      summary: Authenticate with Google ID token
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"